Move verifier out of signature package
